vasdva

+++ L-S/classes/Variant.php
@@ -11,9 +11,9 @@ class Variant {
 
     public string $title;
 
-    public int $pricePerDay = 0;
+    public float $pricePerDay = 0;
     public string $descriptionPerDay;
 
-    public int $priceSingle = 0;
+    public float $priceSingle = 0;
     public string $descriptionSingle;
 }
